// src/app/core/models/student.ts
export interface IStudent {
  id: number;
  name: string;
  category: string;
  domicile: boolean;
  birthCertificate: boolean;
  marksheets: boolean;
  police: boolean;
  passport: boolean;
  declaration: boolean;
  dob: Date;
  fatherName: string;
  motherName: string;
  lastClassScore: number;
}